% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_bands.R
\name{abc}
\alias{abc}
\title{Calculate area between the curves}
\usage{
abc(object)
}
\arguments{
\item{object}{An object to calculate area between the curves.}
}
\value{
The numeric area between the curves.
}
\description{
\code{abc} is used to calculate the area between two confidence band curves.
}
