% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc.R
\name{auc}
\alias{auc}
\title{Compute area under the curve}
\usage{
auc(object)
}
\arguments{
\item{object}{An object of class \code{svmroc}}
}
\value{
The area under the curve.
}
\description{
\code{auc} is used to compute the area under the ROC curve for
an object of class \code{svmroc}.
}
