% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_bands.R
\name{conf_band}
\alias{conf_band}
\title{SVM ROC confidence bands}
\usage{
conf_band(object, num_boot = 1000, gamma = 0.1, x = seq(0.01, 0.99, 0.01))
}
\arguments{
\item{object}{An object of class \code{svmroc}.}

\item{num_boot}{Number of bootstrap replications. Defaults to 1000.}

\item{gamma}{Complement of confidence level, e.g., \code{gamma = 0.1}
will produce 90\% confidence bands. Defaults to 0.1.}

\item{x}{Values used for interpolation. Defaults to \code{seq(0.01, 0.99, 0.01)}.}
}
\value{
An object of class \code{conf_band}, a list with the following components:
  \code{lower}, a vector of values for the lower confidence band; \code{upper}, a
  vectoro of values for the upper confidence band; \code{y}, values of the ROC curve;
  and \code{x}, values used for interpolation.
}
\description{
\code{conf_band} constructs bootstrap confidence bands for the SVM ROC curve from
  an object of class \code{svmroc}
}
