% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc.R
\name{fit_roc}
\alias{fit_roc}
\title{Estimate ROC curve for a fitted classifier using a new testing set}
\usage{
fit_roc(object, new_X, new_A)
}
\arguments{
\item{object}{An object of class \code{svmroc}.}

\item{new_X}{A new X matrix.}

\item{new_A}{A new set of class labels.}
}
\value{
An object of class \code{roc}, a list with components: \code{sens},
  a vector of sensitivity values, and \code{spec}, a vector of specificity values.
}
\description{
\code{fit_roc} estimates the ROC curve for a fitted classifier (an object
of class \code{svmroc}) using a new testing set.
}
