% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmods.R
\name{genmod1}
\alias{genmod1}
\title{First generative model}
\usage{
genmod1(n, beta, variance = 1)
}
\arguments{
\item{n}{The sample size.}

\item{beta}{A vector of parameters with length equal to the number of covariates.}

\item{variance}{The variance of the error term.}
}
\value{
A list with the following components: \code{A}, a vector of class labels; and
  \code{X}, a matrix of covariates.
}
\description{
\code{genmod1} generates data from a simple linear generative model
  for simulation experiments.
}
