% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_bands.R
\name{get_coverage}
\alias{get_coverage}
\title{Determine coverage}
\usage{
get_coverage(conf_band, svmroc, new_X, new_A)
}
\arguments{
\item{conf_band}{An object of class \code{conf_band}.}

\item{svmroc}{An object of class \code{svmroc}.}

\item{new_X}{New X matrix to determine true ROC curve.}

\item{new_A}{New class assignments to determine true ROC curve.}
}
\value{
A list with the following components: \code{coverage}, numeric, 1 if
  the confidence band covers the true ROC curve and 0 if it does not; \code{true_roc},
  a vector of points on the true ROC curve; \code{x}, the x values for the true ROC
  curve (taken from the object \code{conf_band}).
}
\description{
\code{get_coverage} is used to determine if confidence band covers true ROC curve.
}
