% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc.R
\name{opt_weight.conf_band}
\alias{opt_weight.conf_band}
\title{Calculate optimal weight}
\usage{
\method{opt_weight}{conf_band}(object)
}
\arguments{
\item{object}{An object of class \code{conf_band}.}
}
\value{
A list with the following components: \code{opt_sens}, the sensitivity
  at the optimal weight; \code{opt_spec}, the specificity at the optimal weight.
}
\description{
\code{opt_weight.conf_band} computes the optimal point on the ROC curve,
  calculated as the point closest to (0, 1) in Euclidean distance for an
  object of class \code{conf_band}.
}
