% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc.R
\name{opt_weight.svmroc}
\alias{opt_weight.svmroc}
\title{Calculate optimal weight}
\usage{
\method{opt_weight}{svmroc}(object)
}
\arguments{
\item{object}{An object of class \code{svmroc}.}
}
\value{
A list with the following components: \code{weight}, the optimal weight;
  \code{opt_sens}, the sensitivity at the optimal weight; \code{opt_spec}, the
  specificity at the optimal weight.
}
\description{
\code{opt_weight.svmroc} computes the optimal point on the ROC curve, calculated as the point
  closest to (0, 1) in Euclidean distance from an object of class \code{svmroc}.
}
