% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_bands.R
\name{plot.conf_band}
\alias{plot.conf_band}
\title{Plot SVM ROC curve confidence bands}
\usage{
plot.conf_band(object, xlab = "One minus specificity", ylab = "Sensitivity",
  include_opt = T)
}
\arguments{
\item{object}{An object of class \code{conf_band}.}

\item{xlab}{Label for X axis. Defaults to "One minus specificity".}

\item{ylab}{Label for Y axis. Defaults to "Sensitivity".}

\item{include_opt}{Logical. If \code{TRUE}, the optimal point on the ROC curve
(the closest to (0, 1) in Euclidean distance) is marked on the plot.
Defaults to \code{TRUE}.}
}
\description{
\code{plot.conf_band} produces a plot of the ROC curve with confidence bands
  from an object of class conf_band.
}
