% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc.R
\name{plot.svmroc}
\alias{plot.svmroc}
\title{Plot SVM ROC curve}
\usage{
plot.svmroc(object, xlab = "One minus specificity", ylab = "Sensitivity",
  include_opt = T)
}
\arguments{
\item{object}{An object of class \code{svmroc}.}

\item{xlab}{Label for X axis. Defaults to "One minus specificity".}

\item{ylab}{Label for Y axis. Defaults to "Sensitivity".}

\item{include_opt}{Logical. If \code{TRUE}, the optimal point on the ROC curve
(the closest to (0, 1) in Euclidean distance) is marked on the plot.
Defaults to \code{TRUE}.}
}
\value{
A plot as a object of class \code{ggplot}.
}
\description{
\code{plot.svmroc} returns a plot of the ROC curve associated
with an object of class \code{svmroc}.
}
