% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc.R
\name{predict.svmroc}
\alias{predict.svmroc}
\title{Predict method for \code{svmroc}}
\usage{
predict.svmroc(object, newdata, weight = NULL)
}
\arguments{
\item{object}{An object of class \code{svmroc}}

\item{newdata}{New data on which to predict, a matrix with the same columns as
\code{X} used for the fit producing \code{object}.}

\item{weight}{The weight to used for prediction. If \code{NULL}, the optimal weight
is calculated and used. Defaults to \code{NULL}.}
}
\value{
The vector of predicted class assignments.
}
\description{
\code{predict.svmroc} produces predicted class assignments based on the fitted SVM.
}
