% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc.R
\name{svmroc}
\alias{svmroc}
\title{SVM ROC curves}
\usage{
svmroc(X, A, new_X, new_A, kernel = "linear", lambdas = c(0.1, 0.25, 0.5,
  1), sigmas = c(0.1, 0.25, 0.5, 1), num_folds = 10, weights = seq(0.01,
  0.99, 0.01), seed = NULL)
}
\arguments{
\item{X}{A matrix of covariates used for fitting the SVM.}

\item{A}{A factor with two levels or an object that can be coerced to a factor with
two levels; used as the response when fitting the SVM.}

\item{new_X}{A matrix of covariates used as the testing set for estimating the ROC curve.}

\item{new_A}{A factor with two levels or an object that can be coerced to a factor with
two levels; used as the response for the testing set for estimating the ROC curve.
Must have the same levels as \code{Y}.}

\item{kernel}{The kernel used for fitting the SVM, either "linear" or "Gaussian".
Defaults to "linear".}

\item{lambdas}{A numeric vector of penalty parameters to select from. Defaults to
\code{c(0.1, 0.25, 0.5, 1)}.}

\item{sigmas}{A numeric vector of bandwidth parameters to select from. Defaults to
\code{c(0.1, 0.25, 0.5, 1)}.}

\item{num_folds}{Number of folds for cross-validation to select tuning parameters.
Defaults to 10.}

\item{weights}{A vector of weights to use for the weighted SVM. Defaults to
\code{seq(0.01, 0.99, 0.01)}.}

\item{seed}{Random number seed to set. If \code{NULL}, no seed is set. Defaults to \code{NULL}.}
}
\value{
An object of class \code{svmroc}, a list with the following components:
  \code{sens}, estimated sensitivities across weights; \code{spec}, estimated
  specificities across weights; \code{models}, a list of models fit for each weight;
  \code{weights}, the vector of weights used in the fit;
  \code{new_X}, the matrix of covariates in the testing set; and \code{new_Y}, the
  responses in the testing set.
}
\description{
\code{svmroc} fits a weighted support vector machine to estimate a receiver operating
  characteristic curve
}
