% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc_cv.R
\name{svmroc_cv}
\alias{svmroc_cv}
\title{Estimate SVM ROC curve using cross-validation}
\usage{
svmroc_cv(X, A, kernel = "linear", fold_id = NULL, num_folds_roc = 10,
  lambdas = c(0.1, 0.25, 0.5, 1), sigmas = c(0.1, 0.25, 0.5, 1),
  num_folds_params = 10, weights = seq(0.01, 0.99, 0.01), seed = NULL)
}
\arguments{
\item{X}{A matrix of covariates used for fitting the SVM.}

\item{A}{A factor with two levels or an object that can be coerced to a factor with
two levels; used as the response when fitting the SVM.}

\item{kernel}{The kernel used for fitting the SVM, either "linear" or "Gaussian".
Defaults to "linear".}

\item{fold_id}{A vector indicating which fold each observation belows to when estimating
the ROC curve. If \code{NULL}, folds are generated randomly. Defaults to \code{NULL}}

\item{num_folds_roc}{Number of folds used to estimate ROC curve. Ignored if \code{fold_id}
is not \code{NULL}. Defaults to 10.}

\item{lambdas}{A numeric vector of penalty parameters to select from. Defaults to
\code{c(0.1, 0.25, 0.5, 1)}.}

\item{sigmas}{A numeric vector of bandwidth parameters to select from. Defaults to
\code{c(0.1, 0.25, 0.5, 1)}.}

\item{num_folds_params}{Number of folds for cross-validation to select tuning parameters.
Defaults to 10.}

\item{weights}{A vector of weights to use for the weighted SVM. Defaults to
\code{seq(0.01, 0.99, 0.01)}.}

\item{seed}{Random number seed to set. If \code{NULL}, no seed is set. Defaults to \code{NULL}.}
}
\value{
An object of class \code{svmroc}, a list with only the following components:
  \code{sens}, a vector of sensitivities across weights, averaged across folds;
  \code{spec}, a vector of specificities across weights, averaged across folds.
}
\description{
\code{svmroc_cv} estimates the SVM ROC curve using cross-validation.
}
