% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmroc.R
\name{true_roc}
\alias{true_roc}
\title{Estimate ROC curve for fitted classifier with linear interpolation}
\usage{
true_roc(object, new_X, new_A, points)
}
\arguments{
\item{object}{An object of class svmroc}

\item{new_X}{New X matrix to determine true ROC curve.}

\item{new_A}{New class assignments to determine true ROC curve.}

\item{points}{Points at which to interpolate the ROC curve.}
}
\value{
A list with components: \code{x} and \code{y} giving the coordinates of
  the ROC curve.
}
\description{
\code{true_roc} allows for plotting an approximation to the true ROC curve along with
  confidence bands.
}
