% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmods.R
\name{genmod2}
\alias{genmod2}
\title{Second generative model}
\usage{
genmod2(n, mu, beta, variance, q)
}
\arguments{
\item{n}{Sample size.}

\item{mu}{A parameter giving the mean (times plus or minus one).}

\item{beta}{A vector of parameters with length equal to the number of covariates.}

\item{variance}{The variance of the error term.}

\item{q}{A parameter to control the proportion that is diseased.}
}
\value{
A list with the following components: \code{A}, a vector of class labels; and
  \code{X}, a matrix of covariates.
}
\description{
\code{genmod2} generates data from a simple nonlinear generative model for
  simulation experiments.
}
